# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Список всех комментариев",
                "operationId": "getCommentsList",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/commentData"
                                            }
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Создать новый комментарий",
                "operationId": "createComment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/storeCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Комментарий создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/commentData"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/storeCommentValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/comments/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получить комментарий по ID",
                "operationId": "getCommentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/commentData"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Обновить комментарий",
                "operationId": "updateComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updateCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комментарий обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/commentData"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/updateCommentValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Удалить комментарий",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "500Error": {
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "type": "string",
                        "example": "Error string"
                    }
                },
                "type": "object"
            },
            "404Error": {
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "type": "string",
                        "example": "Not found"
                    }
                },
                "type": "object"
            },
            "storeCommentRequest": {
                "description": "Запрос на создание комментария",
                "required": [
                    "post_id",
                    "user_id",
                    "body"
                ],
                "properties": {
                    "post_id": {
                        "description": "Идентификатор поста",
                        "type": "integer",
                        "example": 10
                    },
                    "user_id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "body": {
                        "description": "Текст комментария",
                        "type": "string",
                        "example": "Отличный пост!"
                    }
                },
                "type": "object"
            },
            "storeCommentValidationErrorResponse": {
                "description": "Ответ при ошибках валидации при создании комментария",
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "properties": {
                            "post_id": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The post_id field is required."
                                }
                            },
                            "user_id": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The user_id field is required."
                                }
                            },
                            "body": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The body field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "updateCommentRequest": {
                "description": "Запрос на обновление комментария",
                "required": [
                    "body"
                ],
                "properties": {
                    "body": {
                        "description": "Новый текст комментария",
                        "type": "string",
                        "example": "Обновлённый текст комментария"
                    }
                },
                "type": "object"
            },
            "updateCommentValidationErrorResponse": {
                "description": "Ответ при ошибках валидации при обновлении комментария",
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {},
                        "example": []
                    },
                    "error": {
                        "properties": {
                            "body": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The body field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "commentData": {
                "description": "Данные комментария",
                "required": [
                    "id",
                    "post_id",
                    "user",
                    "body",
                    "created"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор комментария",
                        "type": "integer",
                        "example": 42
                    },
                    "post_id": {
                        "description": "Идентификатор поста",
                        "type": "integer",
                        "example": 10
                    },
                    "user": {
                        "$ref": "#/components/schemas/userData"
                    },
                    "body": {
                        "description": "Текст комментария",
                        "type": "string",
                        "example": "Отличный пост!"
                    },
                    "created": {
                        "description": "Дата и время создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-05T11:15:00Z"
                    }
                },
                "type": "object"
            },
...
            "userData": {
                "description": "Данные пользователя",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Alice"
                    }
                },
                "type": "object"
            }
        }
    },
...
}